Rename to .interleaveOrdered, better place for test
